https

+++ routes/dishRouter.js
@@ -11,6 +11,7 @@ const dishRouter = express.Router();
 
 //we will "mount" this in a short while
 //we are declaring this at a single location 
+var authenticate = require('../authenticate');
 
 dishRouter.use(bodyParser.json());
 
@@ -19,7 +20,10 @@ dishRouter.route('/')
 .get((req,res,next) => {
   //all dishes going to be returned 
   //returns promise 
+  //when dishes are being constructed, we are creating author field 
+  //will be populated
   Dishes.find({})
+  .populate('comments.author')
   .then((dishes) => {
     res.statusCode = 200;
     res.setHeader('Content-Type', 'application/json');
@@ -27,7 +31,8 @@ dishRouter.route('/')
   }, (err) => next(err))
   .catch((err) => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
+  authenticate.verifyAdmin(req, res, next);
   Dishes.create(req.body)
   .then((dish) => {
     console.log('Dish created', dish);
@@ -37,11 +42,12 @@ dishRouter.route('/')
   }, (err) => next(err))
   .catch((err) => next(err));
 })
-.put((req, res, next) => {
+.put(authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -54,6 +60,7 @@ dishRouter.route('/')
 dishRouter.route('/:dishId')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
@@ -61,11 +68,12 @@ dishRouter.route('/:dishId')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
-.put((req, res, next) => {
+.put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body
     }, { new: true })
@@ -76,7 +84,8 @@ dishRouter.route('/:dishId')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp) => {
         res.statusCode = 200;
@@ -89,6 +98,7 @@ dishRouter.route('/:dishId')
 dishRouter.route('/:dishId/comments')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         if (dish != null) {
             res.statusCode = 200;
@@ -103,10 +113,17 @@ dishRouter.route('/:dishId/comments')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
+            //need to reecall that the body 
+            //has the comment but not the author
+            //the user is the author 
+            //by using verifyUser in passport we have
+            //already the information of the user
+            //we just need their id
+            req.body.author = req.user._id;
             dish.comments.push(req.body);
             dish.save()
             .then((dish) => {
@@ -123,12 +140,13 @@ dishRouter.route('/:dishId/comments')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.put((req, res, next) => {
+.put(authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes/'
         + req.params.dishId + '/comments');
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
@@ -154,6 +172,7 @@ dishRouter.route('/:dishId/comments')
 dishRouter.route('/:dishId/comments/:commentId')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
             res.statusCode = 200;
@@ -173,15 +192,29 @@ dishRouter.route('/:dishId/comments/:commentId')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId
         + '/comments/' + req.params.commentId);
 })
-.put((req, res, next) => {
+.put(authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
+        // Compare the String representation of the ids to 
+        //achieve the expected result
+        //Both .author and req.user._id are of type 'object', 
+        //that's why comparing them returns false
+        //(those are different objects in memory).
+        //Whereas if I convert both values to String and compare them, 
+        //I get the expected result:
+
+        if (!String(dish.comments.id(req.params.commentId).author) == String(req.user._id)) {
+        // not authorized to delete the comment
+            err = new Error('You are not authorized to update a comment that is not yours.');
+            err.status = 403;
+            return next(err);
+        }
+        else if (dish != null && dish.comments.id(req.params.commentId) != null) {
             if (req.body.rating) {
                 dish.comments.id(req.params.commentId).rating = req.body.rating;
             }
@@ -208,10 +241,24 @@ dishRouter.route('/:dishId/comments/:commentId')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
+        // Compare the String representation of the ids to 
+        //achieve the expected result
+        //Both .author and req.user._id are of type 'object', 
+        //that's why comparing them returns false
+        //(those are different objects in memory).
+        //Whereas if I convert both values to String and compare them, 
+        //I get the expected result:
+
+        if (!String(dish.comments.id(req.params.commentId).author) == String(req.user._id)) {
+        // not authorized to delete the comment
+            err = new Error('You are not authorized to delete a comment that is not yours.');
+            err.status = 403;
+            return next(err);
+        }
+        else if (dish != null && dish.comments.id(req.params.commentId) != null) {
             dish.comments.id(req.params.commentId).remove();
             dish.save()
             .then((dish) => {
