+++ routes/leaderRouter.js
@@ -11,6 +11,7 @@ const leaderRouter = express.Router();
 
 //we will "mount" this in a short while
 //we are declaring this at a single location 
+var authenticate = require('../authenticate');
 
 leaderRouter.use(bodyParser.json());
 
@@ -27,7 +28,8 @@ leaderRouter.route('/')
   }, (err) => next(err))
   .catch((err) => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
+  authenticate.verifyAdmin(req, res, next);
   Leaders.create(req.body)
   .then((leader) => {
     console.log('Leader created', leader);
@@ -37,11 +39,12 @@ leaderRouter.route('/')
   }, (err) => next(err))
   .catch((err) => next(err));
 })
-.put((req, res, next) => {
+.put(authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end('PUT operation not supported on /leaders');
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Leaders.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -61,11 +64,12 @@ leaderRouter.route('/:leaderId')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end('POST operation not supported on /leaders/'+ req.params.leaderId);
 })
-.put((req, res, next) => {
+.put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Leaders.findByIdAndUpdate(req.params.leaderId, {
         $set: req.body
     }, { new: true })
@@ -76,7 +80,8 @@ leaderRouter.route('/:leaderId')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Leaders.findByIdAndRemove(req.params.leaderId)
     .then((resp) => {
         res.statusCode = 200;
